requirements.txt
Flask==2.0.2
tensorflow==2.8.0
pillow==8.3.2
requests==2.26.0
scikit-learn==0.24.2
transformers==4.11.3
torch==1.9.0
Werkzeug==2.0.2
protobuf==3.20.3
